chore(code-health-gate): Rename "delta" to Code Health Gate
