+++ src/extension.ts
@@ -5,9 +5,9 @@ import { onDidChangeConfiguration } from './configuration';
 import CsDiagnostics from './cs-diagnostics';
 import { CsExtensionState } from './cs-extension-state';
 import { register as registerCsDoc } from './csdoc';
-import { DeltaAnalyser, registerDeltaCommand } from './delta/analyser';
-import { registerDeltaAnalysisDecorations } from './delta/presentation';
-import { DeltaAnalysisView } from './delta/tree-view';
+import { DeltaAnalyser, registerDeltaCommand } from './code-health-gate/analyser';
+import { registerDeltaAnalysisDecorations } from './code-health-gate/presentation';
+import { CodeHealthGateView } from './code-health-gate/tree-view';
 import { ensureLatestCompatibleCliExists } from './download';
 import { reviewDocumentSelector } from './language-support';
 import { outputChannel } from './log';
@@ -76,7 +76,7 @@ function startExtension(context: vscode.ExtensionContext, cliPath: string) {
   context.subscriptions.push(new ReviewExplorerView());
   registerReviewDecorations(context);
 
-  context.subscriptions.push(new DeltaAnalysisView());
+  context.subscriptions.push(new CodeHealthGateView());
   registerDeltaAnalysisDecorations(context);
 
   // Add Review CodeLens support
